+++ internal/provider/provider.go
@@ -129,6 +129,8 @@ func (p *LinkRidgeCloudProvider) Resources(_ context.Context) []func() resource.
 // DataSources returns the list of data source types supported by this provider.
 func (p *LinkRidgeCloudProvider) DataSources(_ context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
+		NewAccountsDataSource,
+		NewQRWorkspacesDataSource,
 		NewServicesDataSource,
 	}
 }